Remove user_location.
Pull out util/geolocation utility.
