+++ setup.cfg
@@ -35,7 +35,6 @@ edx.analytics.tasks =
     sqoop-import = edx.analytics.tasks.sqoop:SqoopImportFromMysql
     dump-student-module = edx.analytics.tasks.database_exports:StudentModulePerCourseTask
     export-student-module = edx.analytics.tasks.database_exports:StudentModulePerCourseAfterImportWorkflow
-    last-country = edx.analytics.tasks.user_location:LastCountryForEachUser
     export-events = edx.analytics.tasks.event_exports:EventExportTask
     export-events-by-course = edx.analytics.tasks.event_exports_by_course:EventExportByCourseTask
     user-activity = edx.analytics.tasks.user_activity:CourseActivityWeeklyTask
